# Fix serialization of nested objects